[x] move overlay feature lint to its own file
This lint has already gotten pretty long and is about to get longer.
